Removing getLabel, and simply returning a simple label for editProfile
