+++ src/registerPanes.js
@@ -4,10 +4,10 @@ module.exports = function registerPanes (register) {
    */
   // Developer designed:
 
-  register(require('./profile/profilePane')) // View someone's public profile - dominates all other panes.
-  register(require('./profile/editProfilePane')) // Edit my profile. App. 201900802
+  register(require('./profile/profile.view')) // View someone's public profile - dominates all other panes.
+  register(require('./profile/editProfile.view')) // Edit my profile. App. 201900802
 
-  register(require('./trustedApplications/trustedApplicationsPane')) // must be registered before basicPreferences
+  register(require('./trustedApplications/trustedApplications.view')) // must be registered before basicPreferences
   register(require('./dashboard/dashboardPane'))
   register(require('./dashboard/basicPreferences')) // 20190702
   register(require('issue-pane'))
